test_dgrid/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6
aiohttp==3.7.4.post0
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-asyncio==0.15.1